Restructure main.py
main.py, at several hundred lines long, was getting ridiculous and out of hands. This introduces a new pattern of only declaring the @kopf.on.* handlers in main.py itself, and implementing the handling in a separate package, which is cleaner.
